Created isSignedIn global state
