+++ src/web/components/Links.tsx
@@ -7,7 +7,12 @@ import { palette } from '@guardian/src-foundations';
 import { textSans } from '@guardian/src-foundations/typography';
 import { from } from '@guardian/src-foundations/mq';
 
-import { SignIn } from '@root/src/web/components/SignIn';
+import { DropdownLinkType, Dropdown } from '@root/src/web/components/Dropdown';
+import ProfileIcon from '@frontend/static/icons/profile.svg';
+
+type Props = {
+    isSignedIn?: boolean;
+};
 
 const search = css`
     :after {
@@ -70,7 +75,7 @@ const linkTablet = ({ showAtTablet }: { showAtTablet: boolean }) => css`
     }
 `;
 
-const seperator = css`
+const seperatorStyles = css`
     border-left: 1px solid ${palette.brand.pastel};
     float: left;
     height: 24px;
@@ -82,7 +87,7 @@ const seperator = css`
     }
 `;
 
-const seperatorHide = css`
+const seperatorHideStyles = css`
     border-left: 1px solid ${palette.brand.pastel};
     float: left;
     height: 24px;
@@ -94,22 +99,27 @@ const seperatorHide = css`
     }
 `;
 
-const Search: React.FC<{
+const Search = ({
+    className,
+    children,
+    href,
+    dataLinkName,
+}: {
     href: string;
     className?: string;
     dataLinkName: string;
-}> = ({ className, children, href, dataLinkName, ...props }) => (
+    children: JSXElements;
+}) => (
     <a
         href={href}
         className={cx(search, className)}
-        {...props}
         data-link-name={dataLinkName}
     >
         {children}
     </a>
 );
 
-const links = css`
+const linksStyles = css`
     position: absolute;
     left: 10px;
     top: 0;
@@ -132,15 +142,54 @@ const links = css`
     }
 `;
 
-const jobsUrl = 'https://jobs.theguardian.com/?INTCMP=jobs_uk_web_newheader';
-const datingUrl =
-    'https://soulmates.theguardian.com/?INTCMP=soulmates_uk_web_newheader';
+const identityLinks: DropdownLinkType[] = [
+    {
+        url: `https://profile.theguardian.com/user/id/123`, // TODO use actual user ID once we have a user model
+        title: 'Comments and replies',
+        dataLinkName: 'nav2 : topbar : comment activity',
+    },
+    {
+        url: `https://profile.theguardian.com/public/edit`,
+        title: 'Public profile',
+        dataLinkName: 'nav2 : topbar : edit profile',
+    },
+    {
+        url: `https://profile.theguardian.com/account/edit`,
+        title: 'Account details',
+        dataLinkName: 'nav2 : topbar : account details',
+    },
+    {
+        url: `https://profile.theguardian.com/email-prefs`,
+        title: 'Emails and marketing',
+        dataLinkName: 'nav2 : topbar : email prefs',
+    },
+    {
+        url: `https://profile.theguardian.com/membership/edit`,
+        title: 'Membership',
+        dataLinkName: 'nav2 : topbar : membership',
+    },
+    {
+        url: `https://profile.theguardian.com/contribution/recurring/edit`,
+        title: 'Contributions',
+        dataLinkName: 'nav2 : topbar : contributions',
+    },
+    {
+        url: `https://profile.theguardian.com/digitalpack/edit`,
+        title: 'Digital pack',
+        dataLinkName: 'nav2 : topbar : subscriptions',
+    },
+    {
+        url: `https://profile.theguardian.com/signout`,
+        title: 'Sign out',
+        dataLinkName: 'nav2 : topbar : sign out',
+    },
+];
 
-export const Links: React.FC = () => (
-    <div className={links}>
-        <div className={seperator} />
+export const Links = ({ isSignedIn }: Props) => (
+    <div id="links-root" className={linksStyles}>
+        <div className={seperatorStyles} />
         <a
-            href={jobsUrl}
+            href="https://jobs.theguardian.com/?INTCMP=jobs_uk_web_newheader"
             className={cx(linkTablet({ showAtTablet: false }), linkStyles)}
             data-link-name="nav2 : job-cta"
         >
@@ -148,22 +197,40 @@ export const Links: React.FC = () => (
         </a>
 
         <a
-            href={datingUrl}
+            href="https://soulmates.theguardian.com/?INTCMP=soulmates_uk_web_newheader"
             className={cx(linkTablet({ showAtTablet: false }), linkStyles)}
             data-link-name="nav2 : soulmates-cta"
         >
             Dating
         </a>
-        <div className={seperatorHide} />
+        <div className={seperatorHideStyles} />
 
-        <SignIn linkStyles={linkStyles} />
+        {isSignedIn ? (
+            <div className={linkStyles}>
+                <ProfileIcon />
+                <Dropdown
+                    label="My account"
+                    links={identityLinks}
+                    id="my-account"
+                    dataLinkName="nav2 : topbar: my account"
+                />
+            </div>
+        ) : (
+            <a
+                className={linkStyles}
+                href="https://profile.theguardian.com/signin?INTCMP=DOTCOM_NEWHEADER_SIGNIN&ABCMP=ab-sign-in"
+                data-link-name="nav2 : topbar : signin"
+            >
+                <ProfileIcon /> Sign in
+            </a>
+        )}
 
         <Search
             className={cx(linkTablet({ showAtTablet: false }), linkStyles)}
             href="https://www.google.co.uk/advanced_search?q=site:www.theguardian.com"
             dataLinkName="nav2 : search"
         >
-            <SearchIcon /> Search
+            <SearchIcon />
         </Search>
     </div>
 );
